requirements.txt
streamlit
pandas
pytrends
deep-translator
schedule
fpdf2
requests
google-genai
python-dotenv
